requirements.txt
astroid==2.4.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
future==0.18.2
idna==2.9
ImageScraper==2.0.7
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.2
pylint==2.5.2
PySocks==1.7.1
requests==2.23.0
requests-oauthlib==1.3.0
setproctitle==1.1.10
SimplePool==0.1
six==1.14.0
soupsieve==2.0.1
toml==0.10.1
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
